pyproject.toml
[project]
name = "image-quality"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.5.2",
    "pillow>=11.1.0",
    "requests>=2.32.3",
    "torch>=2.6.0",
    "transformers>=4.50.2",
]
